Use separate 'write enable' signal for each byte in data word sent to RAM
I am (tentatively) planning to implement the AD-1XX line on a Xilinx
Spartan-6 FPGA. The "Block RAM User's Guide" for these FPGAs explains that
when configured with 32-bit wide data path to/from RAM, their BRAMs have
separate write enable lines for each byte.

This is perfect for implementing RISC-V byte/halfword/word-sized stores.

(It also seems that the BRAMs will have only one write port and one read port
each. This throws a bit of a wrench into things, as I was hoping to have 2
read ports and 1 write port. It means that after reading out an instruction
to execute, we can't load other data from RAM in the same cycle. So loads
will have to run in 2 cycles. That will be worked on next.)